# Add index on clinic_activity_logs.numeric_value for query performance-created-by-agentic