--- a/src/main/resources/db/migration/V2025.05.21__add_numeric_value_index_to_clinic_activity_logs.sql
+++ b/src/main/resources/db/migration/V2025.05.21__add_numeric_value_index_to_clinic_activity_logs.sql
@@ -1,0 +1,2 @@
+-- Add index on numeric_value column to improve query performance
+CREATE INDEX IF NOT EXISTS idx_clinic_activity_logs_numeric_value ON clinic_activity_logs(numeric_value);