This PR addresses a critical query performance issue in the clinic_activity_logs table.

Problem:
- Missing index on numeric_value column causing 4000x performance degradation
- Query latency increased from 653.25μs to 2.67s
- Affected query: SELECT id, activity_type, numeric_value, event_timestamp, status_flag, payload FROM clinic_activity_logs WHERE numeric_value = ?
- Affected endpoint: /api/clinic-activity/query-logs

Changes:
1. Added index on numeric_value column in clinic_activity_logs table
2. Updated schema.sql to include the index definition for new deployments

Performance Impact:
- Before: 2.67s query execution time (sequential scan)
- After: 3.745ms query execution time (index scan)
- Improvement: ~713x faster query execution

The index has been added directly to the database for immediate relief, and this PR ensures the index will be created in new deployments.